Update news.html
I didn't know how to make the picture move locations, but if it could be to the right of all the words that would be best cause there's a lot of empty space. Also, if you know how to make the names of the clubs turn red that would be good. Links all work except "rules" and "rubric", they either go to mail or drive cause that's where they come from, do you know how to make it go straight to a pdf of them?